requirements.txt
PyPDF2
unstructured[pdf]
pandas
transformers
datasets
python-dotenv
pyarrow
chromadb
sentence-transformers
